requirements.txt
numpy~=1.24.3
opencv-python~=4.10.0.84
pymysql~=1.0.2
fastapi~=0.112.2
pydantic~=2.8.2
pillow~=10.4.0
uvicorn~=0.20.0
gym~=0.26.2
matplotlib~=3.7.5
torch~=2.4.1
decorator~=4.4.2
scipy~=1.10.1
pygame~=2.6.1
yaml~=0.2.5
pyyaml~=6.0.2
moviepy~=1.0.3
tqdm~=4.67.1
requests~=2.32.3
six~=1.17.0
scikit-image~=0.21.0